% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annotateChiTestResult}
\alias{annotateChiTestResult}
\title{Annotate significant chimeric transcript test results}
\usage{
annotateChiTestResult(chi_test, gtf, se, cutoff = 0.2)
}
\arguments{
\item{chi_test}{A DESeqDataSet object returned from ChimericDrivenTest().}

\item{gtf}{A GRanges object returned from extractGTF().}

\item{se}{A SummarizedExperiment object.}

\item{cutoff}{Adjusted p-value cutoff to extract the significant results.}
}
\value{
A data.frame with extracted test results, fold change, gene symbol, and exon annotation.
}
\description{
Extract significant chimeric transcripts, compute fold change, match gene symbol and transcript annotation including exon rank.
}
