% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{countAnno}
\alias{countAnno}
\title{Counting reads on the exon}
\usage{
countAnno(
  annotation,
  bamfile,
  isPairedEnd = TRUE,
  strandSpecific = 0,
  allowMultiOverlap = TRUE,
  isLongRead = FALSE,
  nthreads = 1,
  ...
)
}
\arguments{
\item{annotation}{A GRanges object, typically from prepareAnno().}

\item{bamfile}{Character vector of BAM file(s).}

\item{isPairedEnd}{Logical scalar or vector, whether the library is paired-end.}

\item{strandSpecific}{Integer: 0 (unstranded), 1 (forward), or 2 (reverse).}

\item{allowMultiOverlap}{Logical, whether to allow reads to overlap multiple features.}

\item{isLongRead}{Logical, whether this is long-read data (e.g. ONT, PacBio).}

\item{nthreads}{Number of threads to use.}

\item{...}{Additional arguments passed to featureCounts().}
}
\description{
Counting the number of reads that fall into each exon
bin defined in the flattened GTF/GFF object.
}
