% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{diffBinPlot}
\alias{diffBinPlot}
\title{Plot the gene model in the control and reference groups}
\usage{
diffBinPlot(
  count,
  conditions,
  annotation,
  idx,
  labels,
  Title,
  gtf,
  txid,
  chi_test,
  gene_name,
  ...
)
}
\arguments{
\item{count}{A data matrix of bin-based counts，extracted from the output of countAnno\code{\link{countAnno}}}

\item{conditions}{A factor vector indicating the conditions (e.g., control and reference) of each sample}

\item{annotation}{A Granges object for annotated transcripts}

\item{idx}{The index of the gene.}

\item{labels}{A character vector, containing labels for the control and reference groups.}

\item{Title}{The title of plot, e.g., gene name.}

\item{gtf}{Character or GRanges. Path to the annotated GTF generated by \code{\link{gffcompareAnno}}, 
or a GRanges object containing exon-level transcript structures, such as the output of \code{\link{extractGTF}}.}

\item{txid}{The transcript id of a specific transcript for the gene}

\item{chi_test}{DESeq2 object. The output from the ChimericDrivenTest, extracted by \code{\link{extractTest}}}

\item{gene_name}{The gene name.}
}
\description{
Plot the gene model in the control and reference groups
}
