% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractGTF}
\alias{extractGTF}
\title{Extract GTF Based on Type and FPKM Filter}
\usage{
extractGTF(
  combineSE = NULL,
  gtfFile = NULL,
  filter = 1,
  type = c("exon", "transcript", "both")
)
}
\arguments{
\item{combineSE}{Optional. A `SummarizedExperiment` object generated by `stringtieCombine()`,
containing GTF metadata in `metadata(combineSE)$gtf` and an `FPKM` assay.}

\item{gtfFile}{Optional. Path to an annotated GTF file. If provided without `combineSE`,
the function imports the GTF directly and filters only by `type`.}

\item{filter}{A numeric value specifying the minimum FPKM threshold for transcripts to be included in the output. 
Default is 1.}

\item{type}{A character string specifying the type of genetic elements to extract. 
Valid options are "exon", "transcript", or "both". Default is "exon".}
}
\value{
A filtered subset of the GTF metadata, 
including only the specified types of genetic elements that meet the FPKM threshold.
}
\description{
This function filters and extracts GTF information from a `SummarizedExperiment` object 
based on the specified element type (exon, transcript, or both) and an FPKM threshold. 
It is designed to work with `SummarizedExperiment` objects containing GTF metadata and FPKM assay data.
}
