% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifFunction.R
\name{pcmFunction}
\alias{pcmFunction}
\title{Convert Probability Matrix to Position Weight Matrix}
\usage{
pcmFunction(pcm)
}
\arguments{
\item{pcm}{A numeric matrix with 4 rows representing nucleotide probabilities for 'A', 'C', 'G', and 'T'.
Each column represents a position in the motif. The values should be proportions and each column should sum up to 1.}
}
\value{
A PWM object that can be used in further analysis with motif matching functions.
}
\description{
Convert Probability Matrix to Position Weight Matrix
}
