% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{prepareAnno}
\alias{prepareAnno}
\title{Preparing the genome annotation object}
\usage{
prepareAnno(
  gtffile,
  singleGens = TRUE,
  transposon = NULL,
  minoverlap = 10,
  cores = 4
)
}
\arguments{
\item{gtffile}{GTF file.}

\item{singleGens}{Whether to allocate the exon overlapping with two genes to a single gene. Default is TRUE.}

\item{transposon}{A GRanges object with transposon data}

\item{minoverlap}{Minimum overlap for \code{\link[IRanges]{findOverlaps}}. Default is 10.}

\item{cores}{Number of cores used for parallel processing. Default is 1.}
}
\description{
Flatten exon appearing multiple times among different transcripts in GTF file

Flatten exon appearing multiple times among different transcripts in GTF file
}
