% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{stringtieAssembly}
\alias{stringtieAssembly}
\title{R wrapper to Run stringtie}
\usage{
stringtieAssembly(
  bam,
  reference = NULL,
  outfile,
  params = "",
  longRead = FALSE
)
}
\arguments{
\item{bam}{Character, BAM file (sorted by genomic coordinates)}

\item{reference}{Character, reference annotation GTF (optional)}

\item{outfile}{Character, output GTF file name}

\item{params}{Character, extra parameters (default "")}

\item{longRead}{Logical, whether the data is long-read (PacBio/ONT). Default FALSE (short-read).}
}
\description{
Run stringtie for transcriptome assembly
}
