% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ChimericDrivenTest}
\alias{ChimericDrivenTest}
\title{Testing for differential TE-chimeric exon usage}
\usage{
ChimericDrivenTest(
  SEobject,
  condition = NULL,
  annotation = NULL,
  maxit = 100,
  niter = 10,
  quiet = FALSE,
  warning = FALSE,
  filterAllZero = FALSE
)
}
\arguments{
\item{SEobject}{An object of RangedSummarizedExperiment class, out from \bold{countAnno}.}

\item{condition}{Vector, indicating the experimental condition of the samples.}

\item{annotation}{The genome annotation object, out from \bold{prepareAnno}.}

\item{maxit}{Control parameter: maximum number of iterations to allow for convergence when calculating dispersion.}

\item{niter}{Number of times to iterate between estimation of means and estimation of dispersion.}

\item{quiet}{Whether to suppress messages at each step.}

\item{warning}{Whether to print warnings at each step.}

\item{filterAllZero}{Logical, whether to remove all-zero rows before fitting the DESeq2 model.}
}
\description{
Detect to what extent TE-chimeric exon affect the expression of the corresponding transcript
}
