% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ChimericDrivenTest}
\alias{ChimericDrivenTest}
\title{Testing for differential TE-chimeric exon usage}
\usage{
ChimericDrivenTest(
  SEobject,
  condition = NULL,
  annotation = NULL,
  maxit = 100,
  niter = 10,
  quiet = FALSE,
  warning = FALSE
)
}
\arguments{
\item{SEobject}{An object of RangedSummarizedExperiment class, out from \bold{countAnno}.}

\item{condition}{Vector, indicating the experimental condition of the samples.}

\item{annotation}{The genome annotation object, out from \bold{prepareAnno}.}

\item{maxit}{control parameter: maximum number of iterations to allow for convergence when calculating dispersion.}

\item{niter}{whether to print messages at each step}

\item{quiet}{number of times to iterate between estimation of means and estimation of dispersion}

\item{warning}{whether to print warning at each step}
}
\description{
Detect to what extent TE-chimeric exon affect the expression of the corresponding transcript
}
