% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{extractGTF}
\alias{extractGTF}
\title{Extract GTF information}
\usage{
extractGTF(combineSE, filter = 1, type = c("exon", "transcript", "both"))
}
\arguments{
\item{combineSE}{A \code{SummarizedExperiment} object, typically the output from the \code{\link{combineSE}} function.
It should contain GTF metadata in its \code{@metadata$gtf} slot and FPKM values in one of its assays.}

\item{filter}{A numeric value specifying the minimum FPKM threshold for transcripts to be included in the output. 
Default is 1.}

\item{type}{A character string specifying the type of genetic elements to extract. 
Valid options are "exon", "transcript", or "both". Default is "exon".}
}
\value{
A filtered subset of the GTF metadata from the `combineSE` object, 
including only the specified types of genetic elements that meet the FPKM threshold.
}
\description{
This function filters and extracts GTF information from a `SummarizedExperiment` object 
based on the specified element type (exon, transcript, or both) and an FPKM threshold.
}
