% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{extractTest}
\alias{extractTest}
\title{Extract results from the differential TE-bin usage test}
\usage{
extractTest(object, filter = TRUE)
}
\arguments{
\item{object}{An object from \bold{ChimericDrivenTest}.}

\item{filter}{Logical. If \code{TRUE} (default), filter out features with \code{NA} adjusted p-values
and features flagged as structural zeros (\code{allZero}).}
}
\description{
Extract results from the differential TE-bin usage test
}
