% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{gffcompareAnno}
\alias{gffcompareAnno}
\title{R wrapper to run gffcompare}
\usage{
gffcompareAnno(
  reference,
  gtffile,
  outfile,
  params = "",
  cleanup = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{reference}{Use a reference annotation file to guide compare assembly process.}

\item{gtffile}{GTF/GFF file(s) with assembled transcripts.}

\item{outfile}{The name of the final output annotated GTF.}

\item{params}{Other parameters passed to gffcompare.}

\item{cleanup}{Logical. Whether to remove `.tmap` and `.refmap` files generated by gffcompare. Default is TRUE.}

\item{overwrite}{Logical. Whether to overwrite existing output files. Default is FALSE.}
}
\description{
The function to compare one or more GTF/GFF files (the query files)
with a reference annotation using gffcompare. The final annotated GTF is renamed
to the user-specified `outfile`. Tracking, loci and summary files are retained,
while temporary `.tmap` and `.refmap` files are removed by default.
}
