% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifFunction.R
\name{MotifSearch}
\alias{MotifSearch}
\title{Search the motif binding sites through TE-chimeric transcripts}
\usage{
MotifSearch(
  object,
  te,
  pwm,
  minoverlap = 15,
  genome = BSgenome.Mmusculus.UCSC.mm10,
  min.score = 0.9,
  filter = 0
)
}
\arguments{
\item{object}{A Granges object of target TE-chimeric transcripts.}

\item{te}{A Granges object of TE location annotations, typically derived from annotated TE reference.}

\item{pwm}{The pwm matrix for the target motif.}

\item{minoverlap}{A non-negative integer specifying the minimum number for ranges to be considered overlapping. Default is 15.}

\item{genome}{The genome to use, from the BSgenome package (e.g., BSgenome.Mmusculus.UCSC.mm10).}

\item{min.score}{A numeric value specifying the minimum score for a PWM match to be considered significant. Default is 0.9.}

\item{filter}{A non-negative integer specifying the threshold of matches present for a TE-chimeric transcripts. This allows for filtering out transcripts with a low number of significant motif matches, focusing the analysis on more likely candidates. Default is 0.}
}
\description{
Search the motif binding sites through TE-chimeric transcripts
}
