% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{prepareAnno}
\alias{prepareAnno}
\title{Preparing the genome annotation object}
\usage{
prepareAnno(gtffile, singleGens = TRUE, transposon = NULL, minoverlap = 10)
}
\arguments{
\item{gtffile}{GTF file.}

\item{singleGens}{Whether to allocate the exon overlapping with two genes to a single gene. Default is TRUE.}

\item{transposon}{A GRanges object with transposon data}

\item{minoverlap}{Minimum overlap for \code{\link[IRanges]{findOverlaps}}. Default is 10.}
}
\description{
Flatten exon appearing multiple times among different transcripts in GTF file
}
