% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{stringtieAssembly}
\alias{stringtieAssembly}
\title{R wrapper to Run stringtie}
\usage{
stringtieAssembly(bam, reference, gtfFiles, params = "")
}
\arguments{
\item{bam}{Character, a bam file from RNA-Seq reads sorted by genomic location}

\item{reference}{Character, a genome reference to guide transcript assembly}

\item{gtfFiles}{Output GTF files}

\item{params}{Other parameters}
}
\description{
R wrapper to run stringtie, enabling the reconstruction of
transcriptome from RNA-seq reads
}
